feat: 당첨 번호, 보너스 번호 클래스 분리
